Fix matches endpoint and add tests for it
The endpoint used to return multiples of profile-matches pairs

+++ src/Application.kt
@@ -193,8 +193,11 @@ fun Application.module(testing: Boolean = false) {
             call.respondText("Inside $it")
         }
 
-        deleteAllData()  // TODO: Remove for production usage
-        createSampleData()
+        if (!testing) {
+            deleteAllData()  // TODO: Remove for production usage
+            createSampleData()
+
+        }
     }
 }
